% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook_sentiment_calculator.R
\name{clean_text}
\alias{clean_text}
\title{clean_text}
\arguments{
\item{data}{filepath to data}
}
\value{
Text that has been cleaned e.g. special characters and certain punctuation removed.
}
\description{
Facebook: Sentiment Calculator
}
\details{
This function pulls the data from your messages folder created by the facebook_message_pull function.
The sentiment of the data is then calculated and written to a pdf file which has bar charts and a sentiment table.
The sentiments the data pulls are: surprise, disgust, sadness, anticipation, joy, fear, anger, and trust. The data also
pulls the positive and negative percentage of the messages as well.
}
\examples{

library(facebookanalysis)

\dontrun{   clean_text(text)}

}
\keyword{clean_text}
