% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook_heartbeat.R
\name{facebook_heartbeat}
\alias{facebook_heartbeat}
\title{facebook_heartbeat}
\usage{
facebook_heartbeat(folder)
}
\arguments{
\item{folder}{filepath to data}
}
\value{
Sentiment image showing valence of message.
}
\description{
Facebook: Heartbeat Creator
}
\details{
This function pulls the data from your messages folder created by the facebook_message_pull function.
The sentiment of the data is then calculated and a "heartbeat" chart is created. The chart spans the entire length of the
conversation that is present in your messenger data.
}
\examples{

library(facebookanalysis)

 \dontrun{facebook_heartbeat(folder = data)}

}
\keyword{facebook_heartbeat}
