% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook_message_pull.R
\name{facebook_message_pull}
\alias{facebook_message_pull}
\title{facebook_message_pull}
\usage{
facebook_message_pull(folder)
}
\arguments{
\item{data}{filepath to data}
}
\description{
Facebook: Message Pull
}
\details{
This function pulls the data from the Facebook zip folder that is downloaded when you retrieve the contents of your data.
Then the data is placed inside of a message folder
}
\examples{
Load the facebookanalysis library

library(facebookanalysis)

  facebook_message_pull(folder = './facebookmessages/messages/inbox/')

}
\keyword{facebook_message_pull}
