% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook_report.R
\name{facebook_report}
\alias{facebook_report}
\alias{folder}
\title{facebook_report}
\usage{
facebook_report(folder)

facebook_report(folder)
}
\arguments{
\item{folder}{filepath to data}
}
\value{
A report that contains the emotions present in the facebook messages sent between participants. Max of 2                     participants.
}
\description{
Facebook: Report Generator
}
\details{
This function pulls the data from your messages folder created by the facebook_message_pull function.
The sentiment of the data is then calculated and written to a pdf file which has bar charts and a sentiment table.
The sentiments the data pulls are: surprise, disgust, sadness, anticipation, joy, fear, anger, and trust. The data also
pulls the positive and negative percentage of the messages as well.
}
\examples{

library(facebookanalysis)

\dontrun{facebook_report(folder = 'messages')}

}
\keyword{facebook_report}
