% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook_unzip.R
\name{facebook_unzip}
\alias{facebook_unzip}
\title{facebook_unzip}
\usage{
facebook_unzip(path)
}
\arguments{
\item{data}{filepath to data}
}
\value{
This function pulls the data from your downloaded Facebook zip file.
}
\description{
Facebook: Zip folder extractor
}
\details{
This function unzips the Facebook zip folder that is downloaded when you retrieve the contents of your data.
Then the data is placed inside of a facebookmessages folder
}
\examples{
Load the facebookanalysis library

library(facebookanalysis)

  facebook_unzip(path = "./facebook.zip")

}
\keyword{facebook_unzip}
