% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_extractor.R
\name{sentiment_extractor}
\alias{sentiment_extractor}
\title{sentiment_extractor}
\usage{
sentiment_extractor(folder)
}
\arguments{
\item{folder}{filepath to data}
}
\value{
Dataframe with sentiment extracted from each conversation.
}
\description{
Facebook: Message Conversation Extractor
}
\details{
This function pulls the sentiment from the .txt files that contain the text you want to look at. The sentiment is then combined into a data frame with the participant's name and the emotions related to that conversation.
}
\examples{

library(facebookanalysis)

\dontrun{sentiment_extractor(folder = data)}

}
\keyword{sentiment_extractor}
