% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_replace.R
\name{version_replace}
\alias{version_replace}
\alias{version$major}
\title{version_replace}
\usage{
version_replace(major,minor)
}
\arguments{
\item{major}{version of R}

\item{minor}{version of R}
}
\description{
version_replace
}
\examples{


  version_replace(major = version$major,minor = version$minor)

}
\keyword{version_replace}
