% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook_sentiment_calculator.R
\name{facebook_sentiment_calculator}
\alias{facebook_sentiment_calculator}
\title{facebook_sentiment_calculator}
\usage{
facebook_sentiment_calculator(folder)

messages
}
\arguments{
\item{folder}{filepath to data}
}
\value{
sentiment analysis
}
\description{
facebook_sentiment_calculator
}
\examples{

library(facebookanalysis)

\dontrun{facebook_sentiment_calculator(folder = data)}

}
\keyword{facebook_sentiment_calculator}
