% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook_sentiment_calculator.R,
%   R/version_replace.R
\name{version_replace}
\alias{version_replace}
\title{version_replace}
\usage{
version_replace(major, minor)

version_replace(major, minor)
}
\arguments{
\item{major}{major version of R}

\item{minor}{minor version of R}

\item{major}{major version of R}

\item{minor}{minor version of R}
}
\description{
version_replace

version_replace
}
\examples{


  version_replace(major = 'major',minor = 'minor')



  version_replace(major = 'major',minor = 'minor')

}
\keyword{version_replace}
